split bitfield signing metrics
